requirements_test.txt
-r requirements.txt

httpx==0.27.0
pytest==8.0.2
factory-boy==3.3.0